Don't import scipy until needed
Jax 0.3.24 now does lazy imports of scipy, which speeds up 'import jax'
tremendously. So this commit changes dinf to do lazy scipy imports.
We have to jump through a few hoops, but this drops 400-500 ms off the
time that the CLI takes to just show help. 'import jax' is still the
worst offender, taking 350 ms, but this is a massive win!

Part of #195.