Use codecov token
Codecov says that tokens aren't necessary for public repos.
While this is true, it is also problematic. Without a token,
codecov queries the Github Actions API to determine the originating
repository for the upload. Unfortunately, this API is rate limited,
and the limit appears to be exceeded fairly often, which results
in a failure to upload the codecov report:

```
There was an error running the uploader: Error uploading to https://codecov.io: Error: There was an error fetching the storage URL during POST: 404 - {'detail': ErrorDetail(string='Unable to locate build via Github Actions API. Please upload with the Codecov repository upload token to resolve issue.', code='not_found')}
```

By setting the repository token, codecov should skip this step.